[Bugfix][Qwen3-TTS] Preserve speaker encoder mean dtype

+++ vllm_omni/model_executor/models/qwen3_tts/qwen3_tts_talker.py
@@ -1,3 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Omni project
+
 from __future__ import annotations
 
 import copy
@@ -127,7 +130,7 @@ class SqueezeExcitationBlock(nn.Module):
         self.sigmoid = nn.Sigmoid()
 
     def forward(self, hidden_states):
-        hidden_states_mean = hidden_states.mean(dim=2, keepdim=True)
+        hidden_states_mean = hidden_states.mean(dim=2, keepdim=True).to(dtype=hidden_states.dtype)
         hidden_states_mean = self.relu(self.conv1(hidden_states_mean))
         hidden_states_mean = self.sigmoid(self.conv2(hidden_states_mean))
         return hidden_states * hidden_states_mean
